# api/openapi.yaml
openapi: 3.0.3
info:
  title: intern-api
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    V1 API contract for Intern at `intern.bekti.com`.

    Authentication modes:

    - browser traffic arrives behind trusted forward-auth and is normalized by the API
    - public clients authenticate with bearer JWT access tokens minted by the device code flow

    Authorization is enforced server-side. Admin-only operations require membership in the configured admin group set, which defaults to `Super-Users`.
servers:
  - url: https://intern.bekti.com
    description: Production
  - url: http://intern-api:8080
    description: Local Compose network
...
paths:
  /api/v1/profile:
    get:
      operationId: getProfile
      tags: [Profile]
      summary: Get the current authenticated user's profile.
      description: |
        Returns the current user's identity and group membership after the API resolves the caller from either trusted forward-auth headers or a bearer JWT.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/profile/sessions:
    get:
      operationId: listProfileSessions
      tags: [Profile]
      summary: List active auth sessions for the current user.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 25
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: Active sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/profile/sessions/{id}/revoke:
    post:
      operationId: revokeProfileSession
      tags: [Profile]
      summary: Revoke one active session belonging to the current user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session revoked or already absent.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/profile/sessions/revoke_others:
    post:
      operationId: revokeOtherProfileSessions
      tags: [Profile]
      summary: Revoke all other active sessions for the current user.
      description: For bearer-authenticated callers, the current bearer session is kept and other active API sessions are revoked. For forward-auth browser callers, all active API sessions for the user are revoked.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Other sessions revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/networks/vlans:
    get:
      operationId: listVlans
      tags: [Networks]
      summary: List VLAN definitions.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: VLAN list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VlanList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createVlan
      tags: [Networks]
      summary: Create a VLAN definition.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VlanWrite'
      responses:
        '201':
          description: VLAN created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/networks/vlans/{vlan_id}:
    get:
      operationId: getVlan
      tags: [Networks]
      summary: Get one VLAN definition by VLAN number.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VlanId'
      responses:
        '200':
          description: VLAN record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchVlan
      tags: [Networks]
      summary: Update a VLAN definition by VLAN number.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VlanPatch'
      responses:
        '200':
          description: Updated VLAN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteVlan
      tags: [Networks]
      summary: Delete a VLAN definition by VLAN number.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VlanId'
      responses:
        '204':
          description: VLAN deleted or already absent.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/networks/devices:
    get:
      operationId: listNetworkDevices
      tags: [Networks]
      summary: List network devices.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Device list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkDeviceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createNetworkDevice
      tags: [Networks]
      summary: Create a network device.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkDeviceWrite'
      responses:
        '201':
          description: Device created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkDevice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/networks/devices/{id}:
    get:
      operationId: getNetworkDevice
      tags: [Networks]
      summary: Get one network device.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkDevice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchNetworkDevice
      tags: [Networks]
      summary: Update a network device.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkDevicePatch'
      responses:
        '200':
          description: Updated device.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkDevice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteNetworkDevice
      tags: [Networks]
      summary: Delete a network device.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '204':
          description: Device deleted or already absent.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    VlanId:
      name: vlan_id
      in: path
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 4094
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or credentials invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Caller is authenticated but not authorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    Profile:
      type: object
      required: [username, name, email, groups, is_admin]
      properties:
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        groups:
          type: array
          items:
            type: string
        is_admin:
          type: boolean
    AuthSession:
      type: object
      required:
        - id
        - username
        - client_name
        - created_at
        - expires_at
        - idle_expires_at
        - is_current
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        client_name:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
        idle_expires_at:
          type: string
          format: date-time
        is_current:
          type: boolean
    AuthSessionPagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    AuthSessionPage:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuthSession'
        pagination:
          $ref: '#/components/schemas/AuthSessionPagination'
...
    Vlan:
      type: object
      required:
        - name
        - vlan_id
        - description
        - created_at
        - updated_at
      properties:
        name:
          type: string
        vlan_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 4094
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VlanRef:
      type: object
      required: [name, vlan_id]
      properties:
        name:
          type: string
        vlan_id:
          type: integer
          format: int32
    VlanList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Vlan'
    VlanWrite:
      type: object
      required: [name, vlan_id]
      properties:
        name:
          type: string
        vlan_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 4094
        description:
          type: string
          default: ''
    VlanPatch:
      type: object
      properties:
        name:
          type: string
        vlan_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 4094
        description:
          type: string
    NetworkDevice:
      type: object
      required:
        - id
        - mac_address
        - display_name
        - disabled
        - vlan
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        mac_address:
          type: string
          example: aa:bb:cc:dd:ee:ff
        display_name:
          type: string
        disabled:
          type: boolean
        vlan:
          $ref: '#/components/schemas/VlanRef'
        created_by_user_id:
          type: string
          format: uuid
          nullable: true
        updated_by_user_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NetworkDeviceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NetworkDevice'
    NetworkDeviceWrite:
      type: object
      required: [mac_address, display_name, vlan_id]
      properties:
        mac_address:
          type: string
          example: aa:bb:cc:dd:ee:ff
        display_name:
          type: string
        disabled:
          type: boolean
        vlan_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 4094
    NetworkDevicePatch:
      type: object
      properties:
        mac_address:
          type: string
          example: aa:bb:cc:dd:ee:ff
        display_name:
          type: string
        disabled:
          type: boolean
        vlan_id:
          type: integer
          format: int32
          minimum: 1
          maximum: 4094
...